Remove consumers from ClusterEvent and replace date with timestamp
Documents in capped collections cannot be updated if their size would change
by the update, which is exactly what happened with the "consumers" array.

Replaced the logic with a simple timestamp check (only process cluster events
if they are "new", i. e. their timestamp is greater or equal to the current
timestamp). This means that old cluster events won't be processed after startup
even if they haven't been processed before.